// native/src/graphql/__generated__/graphql.ts
export type Maybe<T> = T | null;
// ...
export type Scalars = {
  ID: { input: string; output: string; }
  String: { input: string; output: string; }
  Boolean: { input: boolean; output: boolean; }
  Int: { input: number; output: number; }
  Float: { input: number; output: number; }
  /** An ISO 8601-encoded date */
  ISO8601Date: { input: any; output: any; }
  /** An ISO 8601-encoded datetime */
  ISO8601DateTime: { input: any; output: any; }
  /** Represents untyped JSON */
  JSON: { input: any; output: any; }
  Upload: { input: any; output: any; }
};
// ...
export type Attachment = {
  __typename?: 'Attachment';
  blob: Blob;
  blobId: Scalars['ID']['output'];
  byteSize: Scalars['Int']['output'];
  contentType: Scalars['String']['output'];
  createdAt: Scalars['ISO8601DateTime']['output'];
  filename: Scalars['String']['output'];
  updatedAt: Scalars['ISO8601DateTime']['output'];
  url: Scalars['String']['output'];
};

export type AttributeError = {
  __typename?: 'AttributeError';
  attribute: Scalars['String']['output'];
  errors: Array<Scalars['String']['output']>;
};

export type AudioVariant = {
  __typename?: 'AudioVariant';
  audioFile?: Maybe<Attachment>;
  bitRate?: Maybe<Scalars['Int']['output']>;
  createdAt: Scalars['ISO8601DateTime']['output'];
  digitalRemaster: DigitalRemaster;
  digitalRemasterId: Scalars['ID']['output'];
  format: Scalars['String']['output'];
  id: Scalars['ID']['output'];
  updatedAt: Scalars['ISO8601DateTime']['output'];
};

/** The connection type for AudioVariant. */
export type AudioVariantConnection = {
  __typename?: 'AudioVariantConnection';
  /** A list of edges. */
  edges: Array<AudioVariantEdge>;
  /** A list of nodes. */
  nodes: Array<AudioVariant>;
  /** Information to aid in pagination. */
  pageInfo: PageInfo;
};
// ...
export type Composition = {
  __typename?: 'Composition';
  compositionLyric: CompositionLyric;
  compositionLyrics: CompositionLyricConnection;
  compositionRole: CompositionRole;
  compositionRoles: CompositionRoleConnection;
  id?: Maybe<Scalars['ID']['output']>;
  lyric: Lyric;
  lyrics: LyricConnection;
  recording: Recording;
  recordings: RecordingConnection;
  title?: Maybe<Scalars['String']['output']>;
};
// ...
export type DigitalRemaster = {
  __typename?: 'DigitalRemaster';
  album: Album;
  albumId: Scalars['ID']['output'];
  audioVariant: AudioVariant;
  audioVariants: AudioVariantConnection;
  bpm?: Maybe<Scalars['Int']['output']>;
  createdAt?: Maybe<Scalars['ISO8601Date']['output']>;
  duration?: Maybe<Scalars['Int']['output']>;
  externalId?: Maybe<Scalars['String']['output']>;
  id?: Maybe<Scalars['ID']['output']>;
  peakValue?: Maybe<Scalars['Float']['output']>;
  recording: Recording;
  recordingId: Scalars['ID']['output'];
  remasterAgent?: Maybe<RemasterAgent>;
  remasterAgentId?: Maybe<Scalars['ID']['output']>;
  replayGain?: Maybe<Scalars['Float']['output']>;
  tangoCloudId: Scalars['String']['output'];
  updatedAt?: Maybe<Scalars['ISO8601Date']['output']>;
  waveform: Waveform;
  waveformId: Scalars['ID']['output'];
};


export type DigitalRemasterAudioVariantArgs = {
  id: Scalars['ID']['input'];
};


export type DigitalRemasterAudioVariantsArgs = {
  after?: InputMaybe<Scalars['String']['input']>;
  before?: InputMaybe<Scalars['String']['input']>;
  first?: InputMaybe<Scalars['Int']['input']>;
  last?: InputMaybe<Scalars['Int']['input']>;
};

/** The connection type for DigitalRemaster. */
export type DigitalRemasterConnection = {
  __typename?: 'DigitalRemasterConnection';
  /** A list of edges. */
  edges: Array<DigitalRemasterEdge>;
  /** A list of nodes. */
  nodes: Array<DigitalRemaster>;
  /** Information to aid in pagination. */
  pageInfo: PageInfo;
};
// ...
export type ElRecodoSong = {
  __typename?: 'ElRecodoSong';
  createdAt: Scalars['ISO8601DateTime']['output'];
  date: Scalars['ISO8601Date']['output'];
  disk?: Maybe<Scalars['String']['output']>;
  duration?: Maybe<Scalars['Int']['output']>;
  ertNumber: Scalars['Int']['output'];
  formattedTitle?: Maybe<Scalars['String']['output']>;
  id: Scalars['ID']['output'];
  instrumental: Scalars['Boolean']['output'];
  label?: Maybe<Scalars['String']['output']>;
  lyrics?: Maybe<Scalars['String']['output']>;
  lyricsYear?: Maybe<Scalars['Int']['output']>;
  matrix?: Maybe<Scalars['String']['output']>;
  orchestra: ElRecodoOrchestra;
  orchestraId: Scalars['ID']['output'];
  pageUpdatedAt?: Maybe<Scalars['ISO8601DateTime']['output']>;
  people: ElRecodoPersonConnection;
  person: ElRecodoPerson;
  personRole: ElRecodoPersonRole;
  personRoles: ElRecodoPersonRoleConnection;
  speed?: Maybe<Scalars['Int']['output']>;
  style?: Maybe<Scalars['String']['output']>;
  syncedAt: Scalars['ISO8601DateTime']['output'];
  title: Scalars['String']['output'];
  updatedAt: Scalars['ISO8601DateTime']['output'];
};
// ...
export type Genre = {
  __typename?: 'Genre';
  createdAt: Scalars['ISO8601DateTime']['output'];
  id: Scalars['ID']['output'];
  name: Scalars['String']['output'];
  recording: Recording;
  recordings: RecordingConnection;
  updatedAt: Scalars['ISO8601DateTime']['output'];
};
// ...
export type Item = Recording;
// ...
export type Like = {
  __typename?: 'Like';
  createdAt: Scalars['ISO8601DateTime']['output'];
  id: Scalars['ID']['output'];
  likeableId: Scalars['ID']['output'];
  likeableType: Scalars['String']['output'];
  updatedAt: Scalars['ISO8601DateTime']['output'];
  user: User;
  userId: Scalars['ID']['output'];
};

/** The connection type for Like. */
export type LikeConnection = {
  __typename?: 'LikeConnection';
  /** A list of edges. */
  edges: Array<LikeEdge>;
  /** A list of nodes. */
  nodes: Array<Like>;
  /** Information to aid in pagination. */
  pageInfo: PageInfo;
};
// ...
export type Playlist = {
  __typename?: 'Playlist';
  createdAt: Scalars['ISO8601DateTime']['output'];
  description?: Maybe<Scalars['String']['output']>;
  id: Scalars['ID']['output'];
  image?: Maybe<Attachment>;
  like: Like;
  likes: LikeConnection;
  playlistItem: PlaylistItem;
  playlistItems: PlaylistItemConnection;
  public: Scalars['Boolean']['output'];
  recording: Recording;
  recordings: RecordingConnection;
  share: Share;
  shares: ShareConnection;
  slug?: Maybe<Scalars['String']['output']>;
  subtitle?: Maybe<Scalars['String']['output']>;
  system: Scalars['Boolean']['output'];
  tanda: Tanda;
  tandas: TandaConnection;
  title: Scalars['String']['output'];
  updatedAt: Scalars['ISO8601DateTime']['output'];
  user: User;
  userId: Scalars['ID']['output'];
};
// ...
export type PlaylistConnection = {
  __typename?: 'PlaylistConnection';
  /** A list of edges. */
  edges: Array<PlaylistEdge>;
  /** A list of nodes. */
  nodes: Array<Playlist>;
  /** Information to aid in pagination. */
  pageInfo: PageInfo;
};

/** An edge in a connection. */
export type PlaylistEdge = {
  __typename?: 'PlaylistEdge';
  /** A cursor for use in pagination. */
  cursor: Scalars['String']['output'];
  /** The item at the end of the edge. */
  node?: Maybe<Playlist>;
};

export type PlaylistItem = {
  __typename?: 'PlaylistItem';
  createdAt: Scalars['ISO8601DateTime']['output'];
  id: Scalars['ID']['output'];
  item: Item;
  itemId: Scalars['ID']['output'];
  itemType: Scalars['String']['output'];
  playlist: Playlist;
  playlistId: Scalars['ID']['output'];
  position: Scalars['Int']['output'];
  updatedAt: Scalars['ISO8601DateTime']['output'];
};

/** The connection type for PlaylistItem. */
export type PlaylistItemConnection = {
  __typename?: 'PlaylistItemConnection';
  /** A list of edges. */
  edges: Array<PlaylistItemEdge>;
  /** A list of nodes. */
  nodes: Array<PlaylistItem>;
  /** Information to aid in pagination. */
  pageInfo: PageInfo;
};
// ...
export type Recording = {
  __typename?: 'Recording';
  audioVariant: AudioVariant;
  audioVariants: AudioVariantConnection;
  composition?: Maybe<Composition>;
  compositionId?: Maybe<Scalars['ID']['output']>;
  digitalRemaster: DigitalRemaster;
  digitalRemasters: DigitalRemasterConnection;
  elRecodoSong?: Maybe<ElRecodoSong>;
  elRecodoSongId?: Maybe<Scalars['ID']['output']>;
  genre?: Maybe<Genre>;
  genreId?: Maybe<Scalars['ID']['output']>;
  id: Scalars['ID']['output'];
  like: Like;
  likedByCurrentUser?: Maybe<Scalars['Boolean']['output']>;
  likes: LikeConnection;
  lyric: Lyric;
  lyrics: LyricConnection;
  orchestra?: Maybe<Orchestra>;
  orchestraId?: Maybe<Scalars['ID']['output']>;
  playback: Playback;
  playbacks: PlaybackConnection;
  playbacksCount: Scalars['Int']['output'];
  playlistItem: PlaylistItem;
  playlistItems: PlaylistItemConnection;
  recordLabel?: Maybe<RecordLabel>;
  recordLabelId?: Maybe<Scalars['ID']['output']>;
  recordedDate?: Maybe<Scalars['ISO8601Date']['output']>;
  recordingSinger: RecordingSinger;
  recordingSingers: RecordingSingerConnection;
  recordingType: RecordingTypeEnum;
  share: Share;
  shares: ShareConnection;
  singer: Singer;
  singers: SingerConnection;
  slug: Scalars['String']['output'];
  tag: Tag;
  tagging: Tagging;
  taggings: TaggingConnection;
  tags: TagConnection;
  tanda: Tanda;
  tandaRecording: TandaRecording;
  tandaRecordings: TandaRecordingConnection;
  tandas: TandaConnection;
  timePeriod?: Maybe<TimePeriod>;
  timePeriodId?: Maybe<Scalars['ID']['output']>;
  waveform: Waveform;
  waveforms: WaveformConnection;
  year?: Maybe<Scalars['Int']['output']>;
};
// ...
export type RecordingConnection = {
  __typename?: 'RecordingConnection';
  /** A list of edges. */
  edges: Array<RecordingEdge>;
  /** A list of nodes. */
  nodes: Array<Recording>;
  /** Information to aid in pagination. */
  pageInfo: PageInfo;
};
// ...
export enum RecordingTypeEnum {
  Live = 'live',
  Studio = 'studio'
}
// ...
export type Share = {
  __typename?: 'Share';
  createdAt: Scalars['ISO8601DateTime']['output'];
  id: Scalars['ID']['output'];
  updatedAt: Scalars['ISO8601DateTime']['output'];
  user: User;
};

/** The connection type for Share. */
export type ShareConnection = {
  __typename?: 'ShareConnection';
  /** A list of edges. */
  edges: Array<ShareEdge>;
  /** A list of nodes. */
  nodes: Array<Share>;
  /** Information to aid in pagination. */
  pageInfo: PageInfo;
};
// ...
export type Tanda = {
  __typename?: 'Tanda';
  createdAt: Scalars['ISO8601DateTime']['output'];
  description?: Maybe<Scalars['String']['output']>;
  id: Scalars['ID']['output'];
  like: Like;
  likes: LikeConnection;
  playlist: Playlist;
  playlistItem: PlaylistItem;
  playlistItems: PlaylistItemConnection;
  playlists: PlaylistConnection;
  public: Scalars['Boolean']['output'];
  recording: Recording;
  recordings: RecordingConnection;
  share: Share;
  shares: ShareConnection;
  subtitle?: Maybe<Scalars['String']['output']>;
  tandaRecording: TandaRecording;
  tandaRecordings: TandaRecordingConnection;
  title: Scalars['String']['output'];
  updatedAt: Scalars['ISO8601DateTime']['output'];
  user: User;
  userId: Scalars['ID']['output'];
};
// ...
export type TandaConnection = {
  __typename?: 'TandaConnection';
  /** A list of edges. */
  edges: Array<TandaEdge>;
  /** A list of nodes. */
  nodes: Array<Tanda>;
  /** Information to aid in pagination. */
  pageInfo: PageInfo;
};

/** An edge in a connection. */
export type TandaEdge = {
  __typename?: 'TandaEdge';
  /** A cursor for use in pagination. */
  cursor: Scalars['String']['output'];
  /** The item at the end of the edge. */
  node?: Maybe<Tanda>;
};

export type TandaRecording = {
  __typename?: 'TandaRecording';
  createdAt: Scalars['ISO8601DateTime']['output'];
  id: Scalars['ID']['output'];
  position: Scalars['Int']['output'];
  recording: Recording;
  recordingId: Scalars['ID']['output'];
  tanda: Tanda;
  tandaId: Scalars['ID']['output'];
  updatedAt: Scalars['ISO8601DateTime']['output'];
};

/** The connection type for TandaRecording. */
export type TandaRecordingConnection = {
  __typename?: 'TandaRecordingConnection';
  /** A list of edges. */
  edges: Array<TandaRecordingEdge>;
  /** A list of nodes. */
  nodes: Array<TandaRecording>;
  /** Information to aid in pagination. */
  pageInfo: PageInfo;
};
// ...
export type User = {
  __typename?: 'User';
  admin: Scalars['Boolean']['output'];
  createdAt: Scalars['ISO8601DateTime']['output'];
  email: Scalars['String']['output'];
  id: Scalars['ID']['output'];
  like: Like;
  likedRecording: Recording;
  likedRecordings: RecordingConnection;
  likes: LikeConnection;
  playback: Playback;
  playbacks: PlaybackConnection;
  playlist: Playlist;
  playlistItem: PlaylistItem;
  playlistItems: PlaylistItemConnection;
  playlists: PlaylistConnection;
  provider?: Maybe<Scalars['String']['output']>;
  share: Share;
  sharedOrchestra: Share;
  sharedOrchestras: ShareConnection;
  sharedPlaylist: Share;
  sharedPlaylists: ShareConnection;
  sharedRecording: Share;
  sharedRecordings: ShareConnection;
  sharedTanda: Share;
  sharedTandas: ShareConnection;
  shares: ShareConnection;
  tag: Tag;
  tagging: Tagging;
  taggings: TaggingConnection;
  tags: TagConnection;
  tanda: Tanda;
  tandas: TandaConnection;
  uid?: Maybe<Scalars['String']['output']>;
  updatedAt: Scalars['ISO8601DateTime']['output'];
  userPreference?: Maybe<UserPreference>;
  username?: Maybe<Scalars['String']['output']>;
};